test_mwss_index in place.
